Handle successful/failed HTTP responses
If the request is successful, its body will be parsed as JSON and returned.
Otherwise, a `Net::HTTPError` is raised containing the error text and
`Net::HTTPResponse` object.

`WebMock` is used to specify responses. For sake of simple implementation and
consistent tests, it's used in the request generator tests as well, returning a
simple JSON object to check. All real web requests are disabled by
`WebMock.disable_net_connect!`, which lets us verify the proper URL is
generated, as well as prevents hits to any real server. The URL stubs are
repetetive, but nice and explicit.